Change how UPDATE-SHAPE works.
UPDATE-SHAPE now receives a hash table of initargs to set, which each method modifies. This prevents massive amounts of reinitialization calls when many update methods are inherited and override each other's properties, resulting in much faster updating. The only downside is that we need to cons up a hash table.

We now also call UPDATE-SHAPE after REALIZE-RENDERABLE to ensure changes are applied and noticed immediately, which is especially important when the sizing of renderables depends on that of shapes. Previously this would result in the look of a renderable updating, but not the associated bounding in the layout.

Another improvement would be to cache the hash table and check whether the initargs are the same as before, and elide the update if so.